pyproject.toml
[tool.poetry]
name = "rssfeed"
version = "0.1.0"
description = "API for consuming and indexing RSS feeds"
authors = ["Rami <rami.awar.ra@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
sqlmodel = "^0.0.9"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
mypy = "^1.6.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
feedparser = "^6.0.10"
celery = "5.3"
redis = "^5.0.1"
xxhash = "^3.4.1"
requests = "^2.31.0"
flower = "^2.0.1"


[tool.poetry.group.dev.dependencies]
types-python-jose = "^3.3.4.8"
types-passlib = "^1.7.7.13"
ipython = "^8.16.1"
celery-types = "^0.20.0"
types-redis = "^4.6.0.8"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
